Fix issues when pushing sub stack to a different repo
Summary
Pushing to a substack (ie. a branch that is not directly merging to
master/main) is not working accross repository as there is no way to
create a branch in `main_repo` where `branch_B` would merge in
`branch_A`.

So instead create a PR for the forked repository, it's not an ideal
situation but at least it allows to run tests (mostly) and potentially
to have proper discussion on the PR.